+++ cdk/bin/cdk.ts
@@ -2,7 +2,6 @@ import 'source-map-support/register';
 import { App } from 'aws-cdk-lib';
 import type { AdminConsoleProps } from '../lib/admin-console';
 import { AdminConsole } from '../lib/admin-console';
-import { AdminConsoleDynamo } from '../lib/admin-console-dynamo';
 
 const app = new App();
 
@@ -19,6 +18,3 @@ export const prodProps: AdminConsoleProps = {
   domainName: 'support.gutools.co.uk',
 };
 new AdminConsole(app, 'AdminConsole-PROD', prodProps);
-
-new AdminConsoleDynamo(app, 'AdminConsoleDynamo-CODE', { stack: 'support', stage: 'CODE' });
-new AdminConsoleDynamo(app, 'AdminConsoleDynamo-PROD', { stack: 'support', stage: 'PROD' });
