--- cdk/lib/admin-console-dynamo.test.ts
@@ -1,15 +1,0 @@
-import { App } from 'aws-cdk-lib';
-import { Template } from 'aws-cdk-lib/assertions';
-import { AdminConsoleDynamo } from './admin-console-dynamo';
-
-describe('The AdminConsole stack', () => {
-  it('matches the snapshot', () => {
-    const app = new App();
-    const stack = new AdminConsoleDynamo(app, 'AdminConsoleDynamo', {
-      stack: 'support',
-      stage: 'TEST',
-    });
-    const template = Template.fromStack(stack);
-    expect(template.toJSON()).toMatchSnapshot();
-  });
-});
